backend/Leave_Management_Microservice/requirements.txt
uvicorn
fastapi
sqlalchemy
mysql-connector-python
python-jose
python-multipart
requests